clean up pipeline scripts and allow usage of semantic versioning

+++ src/BuildForRelease.ps1
@@ -6,50 +6,47 @@
 
 <#
 .DESCRIPTION
-This script builds the solution for debugging (manual or automatic testing).
+This script builds the solution for debugging (manual or automatic testing) or cleans up the previous build.
 #>
 
 # Set error action preference to stop on errors.
 $ErrorActionPreference = "Stop"
 
+# Import necessary functions from the Common module.
 Import-Module (Join-Path $PSScriptRoot Common.psm1) -Function AssertDotnet, GetArtefactsDir
 
-function CopyContentForDemo($Destination)
-{
+function CopyContentForDemo($Destination) {
     $contentForDemoDir = Join-Path (Split-Path -Parent $PSScriptRoot) "content-for-demo"
-    if (!(Test-Path $contentForDemoDir))
-    {
+    if (!(Test-Path $contentForDemoDir)) {
         throw "The directory with the content for demo does not exist: $contentForDemoDir"
     }
 
     Write-Host "Copying content for demo from $contentForDemoDir to: $Destination"
-
     Get-ChildItem -Path $contentForDemoDir | Copy-Item -Destination $Destination -Recurse -Container -Force
 }
 
-function Main
-{
+function Main {
     # Change to the script root directory.
     Set-Location $PSScriptRoot
 
     # Define the base build directory.
     $baseBuildDir = Join-Path $(GetArtefactsDir) "build" | Join-Path -ChildPath "Release"
 
-    if ($clean)
-    {
+    if ($clean) {
         # Clean up previous build if requested.
-        Write-Host "dotnet clean'ing ..."
+        Write-Host "Cleaning up previous build..."
         dotnet clean
-        if ($LASTEXITCODE -ne 0)
-        {
-            throw "Failed to dotnet clean."
+        if ($LASTEXITCODE -ne 0) {
+            throw "Failed to clean the project."
         }
 
         Write-Host "Removing the build directory: $baseBuildDir"
-        Remove-Item -LiteralPath $baseBuildDir -Force -Recurse
-    }
-    else
-    {
+        if (Test-Path $baseBuildDir) {
+            Remove-Item -LiteralPath $baseBuildDir -Force -Recurse
+        } else {
+            Write-Host "Build directory does not exist: $baseBuildDir"
+        }
+    } else {
         # Ensure dotnet is installed.
         AssertDotnet
 
@@ -59,16 +56,13 @@ function Main
             "AasxServerAspNetCore"
         )
 
-        foreach ($target in $dotnetTargets)
-        {
+        foreach ($target in $dotnetTargets) {
             $buildDir = Join-Path $baseBuildDir $target
 
-            Write-Host "Publishing with dotnet $target to: $buildDir"
-
+            Write-Host "Publishing $target to: $buildDir"
             dotnet publish -c Release -o $buildDir $target
-            if ($LASTEXITCODE -ne 0)
-            {
-                throw "Failed to dotnet publish: $target"
+            if ($LASTEXITCODE -ne 0) {
+                throw "Failed to publish the project: $target"
             }
 
             # Copy content for demo to the build directory.
@@ -79,11 +73,8 @@ function Main
 
 # Store the current location, execute the main function, and return to the original location.
 $previousLocation = Get-Location
-try
-{
+try {
     Main
-}
-finally
-{
+} finally {
     Set-Location $previousLocation
 }
